pyproject.toml
[project]
name = "trawl"
version = "0.1.0"
description = "On-Premise AI Powered Research Assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Udaykumar Dhokia", email = "udaykumardhokia@gmail.com"},
]
maintainers = [
    {name = "Udaykumar Dhokia", email = "udaykumardhokia@gmail.com"},
]
keywords = ["ai", "research", "assistant", "terminal", "tui", "fastapi", "ollama", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Terminals",
]
dependencies = [
    "certifi>=2026.2.25",
    "fastapi[standard]>=0.135.1",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "numpy>=2.4.3",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pyperclip>=1.11.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.48",
    "textual>=8.1.1",
    "trafilatura>=2.0.0",
    "truststore>=0.10.4",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/udaykumar-dhokia/trawl"
Documentation = "https://github.com/udaykumar-dhokia/trawl#readme"
Repository = "https://github.com/udaykumar-dhokia/trawl"
Issues = "https://github.com/udaykumar-dhokia/trawl/issues"
Changelog = "https://github.com/udaykumar-dhokia/trawl/blob/main/CHANGELOG.md"

[project.scripts]
trawl = "trawl.cli:main"
trawl-tui = "trawl.tui_app:run_tui"
trawl-api = "trawl.main:run_api"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trawl"]

[tool.hatch.version]
path = "trawl/__init__.py"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"] # assert false

[tool.ruff.lint.isort]
known-first-party = ["trawl"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sentence_transformers.*",
    "langchain.*",
    "langchain_ollama.*",
    "langchain_google_genai.*",
    "google.generativeai.*",
    "textual.*",
    "trafilatura.*",
    "pgvector.*",
]
ignore_missing_imports = true
